:sparkles: Add python script to create ibus tables
using an IBus table is currently the preferred IME for linux users. Having it here will make it easier to be updated.
This commit adds a python script that automatically converts the AHK
scripts to simple ibus tables, based on the previous most used one for
sitelen pona. The only pip dep needed is uuid.